Rename files based on setuptools `src` layout.
Documentation here: https://setuptools.readthedocs.io/en/latest/userguide/declarative_config.html#using-a-src-layout
