+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = gdscript_docs_maker
-version = 1.3.0
+version = 1.5.1
 description = Create documentation and class references from your Godot GDScript code.
 long_description = file: README.md, CHANGELOG.md
 long_description_content_type = text/markdown
@@ -20,14 +20,21 @@ classifiers =
 
 [options]
 package_dir=
-    =gdscript_docs_maker
-packages = find:
+    =src
+packages=find:
 include_package_data = True
 python_requires = >=3.7
-entry_points = file: gdscript_docs_maker.__main__:main
+entry_points = file: gdscript_docs_maker.modules.__main__:main
 
 [options.packages.find]
-where=gdscript_docs_maker
+where=src
+
+[options.package_data]
+* = *.toml
+
+[options.data_files]
+data = src/gdscript_docs_maker/data/hugo_front_matter.toml
 
 [mypy]
-mypy_path = ./gdscript_docs_maker
+mypy_path = ./src
+
