+++ src/gdscript_docs_maker/config.py
