+++ src/gdscript_docs_maker/gdscript_objects.py
